requirements.txt
OctoPrint
RPi.GPIO>=0.7
OctoPrint_PSUControl
requests
